fix(Discover): String resources fixed
