+++ app/src/main/res/values/strings.xml
@@ -223,8 +223,9 @@
     <string name="ptype_placeholder">Property type</string>
     <string name="approval_placeholder">approval</string>
     <string name="apply_placeholder">Apply filters</string>
-    <string name="cancel_placeholder">Cancel</string>
+    <string name="clean_placeholder">Clean</string>
     <string name="any_placeholder">Any</string>
+    <string name="close">Close</string>
 
     <!-- CTASection -->
     <string name="cta_title">Ready for your next smart adventure?</string>
